Break three persistence import cycles via leaf extraction
Resolves three isolated strongly-connected components flagged by
`lint:circular`, each an instance of the same shape — a "hub" module and its
satellites both importing back from the hub:

1. Usage aggregation
   llm-usage-store ↔ usage-buckets ↔ usage-grouped-buckets. The bucket helpers
   imported UsageDayBucket / UsageTimeRange / UsageBucketOptions /
   GroupByDimension back from llm-usage-store (which imports the helpers).
   Moved those to a new leaf `usage-types.ts`; the helpers now import from the
   leaf and llm-usage-store re-exports the symbols so its public surface is
   unchanged.

2. LLM request log source
   llm-request-log-source ↔ -local / -clickhouse. The two implementations
   imported the `LlmRequestLogSource` interface back from the resolver (whose
   factory dynamically imports them). Moved the interface to a new leaf
   `llm-request-log-source-types.ts`; the resolver re-exports it.

3. Embedding backend
   embedding-backend → embedding-local → embedding-runtime-manager →
   embedding-backend. The runtime manager reached back into the backend to
   clear sticky local-failure state after a background runtime download.
   Inverted that edge through a new leaf `embedding-local-retry.ts`: the
   backend registers a reset hook at load time and the runtime manager invokes
   it. A null hook is a safe no-op (the backend can only have failed if it was
   loaded, which registers the hook), so behavior is preserved.

madge circular chains: 420 → 415; no cycle now touches any of the three
clusters. Public surfaces are unchanged (types re-exported from their original
modules), so no external consumers were modified.
