+++ assistant/src/persistence/llm-usage-store.ts
@@ -28,6 +28,23 @@ import {
   type UsageGroupedBucketRow,
   type UsageGroupedSeriesBucket,
 } from "./usage-grouped-buckets.js";
+import {
+  type GroupByDimension,
+  USAGE_GROUP_BY_DIMENSIONS,
+  type UsageBucketOptions,
+  type UsageDayBucket,
+  type UsageTimeRange,
+} from "./usage-types.js";
+
+// Re-exported so the aggregation module's public surface is unchanged for the
+// route/CLI/test consumers that import these types from here.
+export {
+  type GroupByDimension,
+  USAGE_GROUP_BY_DIMENSIONS,
+  type UsageBucketOptions,
+  type UsageDayBucket,
+  type UsageTimeRange,
+};
 
 // ---------------------------------------------------------------------------
 // Write
@@ -292,12 +309,6 @@ export function queryUnreportedUsageEvents(
 // Aggregation — time-range queries for the usage dashboard
 // ---------------------------------------------------------------------------
 
-/** Epoch-millis time range (inclusive on both ends). */
-export interface UsageTimeRange {
-  from: number;
-  to: number;
-}
-
 export type UsageAggregationFilter = ScheduleAttributionFilter;
 
 /** Aggregate totals across a time range. */
@@ -314,34 +325,6 @@ export interface UsageTotals {
 }
 
 export type UsageGranularity = "daily" | "hourly";
-
-/** A single time bucket with its aggregate totals. */
-export interface UsageDayBucket {
-  /**
-   * Stable unique identifier for the bucket. Safe for use as a SwiftUI/React
-   * list key. Distinct even for DST fall-back duplicate hours (which share the
-   * same `date` string). Daily buckets use `date` directly; hourly buckets use
-   * "YYYY-MM-DD HH:00|<offsetMinutes>" to disambiguate repeated local hours.
-   */
-  bucketId: string;
-  /**
-   * Local-time bucket key in the requested tz:
-   * "YYYY-MM-DD" (daily) or "YYYY-MM-DD HH:00" (hourly).
-   * NOT unique: on DST fall-back days, two 01:00 hourly buckets share this key.
-   * Use `bucketId` as a list identifier and `date` for display/sort only.
-   */
-  date: string;
-  /**
-   * Human-readable label for the bucket, formatted in the requested tz.
-   * Hourly: "3pm". Daily: "Apr 11".
-   */
-  displayLabel?: string;
-  /** Direct input tokens only; cache traffic is tracked separately in totals. */
-  totalInputTokens: number;
-  totalOutputTokens: number;
-  totalEstimatedCostUsd: number;
-  eventCount: number;
-}
 
 /** A grouped breakdown row. */
 export interface UsageGroupBreakdown {
@@ -640,16 +623,6 @@ function fetchRawBucketRows(
   );
 }
 
-/** Options for bucket aggregation. */
-export interface UsageBucketOptions {
-  /**
-   * When true, emit a zero-value bucket for every day (or hour) in the range
-   * even if no events fall inside it. Defaults to false so the CLI and other
-   * callers only see active periods; the chart route opts in.
-   */
-  fillEmpty?: boolean;
-}
-
 /**
  * Return per-day aggregates within the given time range, keyed by local date
  * in the requested timezone (default UTC).
@@ -687,18 +660,6 @@ export function getUsageHourBuckets(
   const rows = fetchRawBucketRows(range, filter);
   return bucketEventsByHour(rows, range, tz, options);
 }
-
-export const USAGE_GROUP_BY_DIMENSIONS = [
-  "actor",
-  "provider",
-  "model",
-  "conversation",
-  "call_site",
-  "inference_profile",
-  "schedule",
-] as const;
-
-export type GroupByDimension = (typeof USAGE_GROUP_BY_DIMENSIONS)[number];
 
 export const USAGE_SERIES_GROUP_BY_DIMENSIONS = [
   "actor",
